stable sort update
branches are now sorting correctly, been through a bit of an issue on the ArgsHashed.__lt__ method, types apparently are not comparable by default, I added a naïve sort between them by comparing their names. I'm not sure this will stay stable in the future or if it's the right ay to do it, but it seems to work.
What could go wrong ? ^^'